+++ History.md
@@ -4,7 +4,7 @@
   isolation.
 * Removed support for konacha_config.js and Konacha.mochaOptions in favor of
   Mocha's own configuration methods. See the README for update instructions.
-* Update mocha (1.5.0) and chai (1.3.0)
+* Update mocha (1.6.0) and chai (1.3.0)
 
 # 1.x-stable